tests: split tests by backend implementations
